Set the preferred interface address for client use
In a multi-interface multi-network environment, it is possible the
private-address selected by juju will not be the preferred one. This
can lead to communication failures between the client charm and the
controller charm.

This change leverages the configuration setting for multicast-intf to
select the preferred address to send over the client relationship.
If multicast-intf is not set, the charm will check with network-get
for network space configuration. Thus guaranteeing the client will
use the correct network for communication.