Guarantee flush of IP for underlying bridge int
When ifdown is called on an interface without an 'addr' setting
there is no guarantee the IP will be flushed from the interface.
If the IP remains on the underlying interface the br-ext bridge
relies on the CPlane SDN fails to respond.

Guarantee the IP is flushed from the underlying bridge interface by
calling a gratuitous `ip flush`